[Fix] 유틸 분리, 주석 제거
